fix:button change and nft information fix
