+++ src/pages/Main/Main.tsx
@@ -9,7 +9,6 @@ import ActionCards from "@/components/main/ActionCards";
 import MainMyAssetInfo from "@/components/main/MainMyAssetInfo";
 import { WelcomeModal } from "@/components/main/Modal/WelcomeModal";
 import MyTokens from "@/components/main/MyTokens";
-import StakeView from "@/components/main/StakeView/StakeView";
 import { useManageReferral } from "@/hooks/api/referral/useManageReferral";
 import { useTrackReferral } from "@/hooks/api/referral/useTrackReferral";
 import { useStakeInfo } from "@/hooks/api/useStakeInfo";
@@ -20,7 +19,6 @@ import "react-toastify/dist/ReactToastify.css";
 const tele = (window as any).Telegram.WebApp;
 
 const Main = () => {
-  const location = useLocation();
   const { address, balance, refreshTonData, connected, tonConnectUI } = useTonConnect();
   const { nftList, isLoading, isError } = useStakeInfo(address);
 
@@ -120,27 +118,6 @@ const Main = () => {
 
     trackReferral();
   }, [trigger, triggerManageReferral]);
-
-  // Show toast message when the user has successfully staked
-  useEffect(() => {
-    const { state } = location;
-
-    if (state?.isStakeSuccess) {
-      toast(`Transaction approved! Your balance will be updated within the next 30 seconds.`, {
-        position: "top-center",
-        autoClose: 5000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-        theme: "light",
-        transition: Slide,
-      });
-
-      history.replaceState(null, "");
-    }
-  }, [location]);
 
   // Calculate the total amount staked
   const totalStaked = useMemo(() => {
